Some refactoring of command-line things. The command line can now be invoked through a package.json script
