+++ package.json
@@ -5,7 +5,6 @@
   "keywords": [
     "couchdb"
   ],
-  "main": "./dist/index.js",
   "repository": "https://github.com/crkn-rcdr/kivik",
   "contributors": [
     {
@@ -20,14 +19,17 @@
     }
   ],
   "license": "Unlicense",
+  "main": "./dist/index.js",
   "bin": {
-    "kivik": "./dist/cli/index.js"
+    "kivik": "./dist/bin.js"
   },
   "engines": {
     "node": ">=12.0.0"
   },
   "scripts": {
-    "build": "rm -rf dist && tsc",
+    "build": "rm -rf dist/** && tsc",
+    "kivik": "run-s build \"kivik:run -- {@}\" --",
+    "kivik:run": "cd example && node ../dist/bin.js",
     "test": "ava -v",
     "watch": "ava -v -w"
   },
@@ -46,7 +48,6 @@
     "yargs": "^16.2.0"
   },
   "devDependencies": {
-    "@ava/typescript": "^1.1.1",
     "@types/dockerode": "^3.2.3",
     "@types/fs-extra": "^9.0.10",
     "@types/node": "^14.14.37",
@@ -71,6 +72,9 @@
     ],
     "require": [
       "ts-node/register"
+    ],
+    "ignoredByWatcher": [
+      "{dist,example,oldtests}/**"
     ]
   }
 }
