+++ src/cli/fixtures.ts
@@ -1,15 +1,15 @@
-import { InitContext } from "../context";
+import { UnloggedContext } from "../context";
 import { createKivikFromContext } from "../kivik";
-import { CommonArgv } from "./parse";
+import { CommonArgv } from ".";
 
-export default (context: InitContext) => {
+export default (unloggedContext: UnloggedContext) => {
 	return {
 		command: "fixtures",
 		describe: "Test that each database's fixtures validate",
 		handler: async (argv: CommonArgv) => {
-			const fullContext = context.withArgv(argv);
+			const context = unloggedContext.withArgv(argv);
 
-			const kivik = await createKivikFromContext(fullContext, "fixtures");
+			const kivik = await createKivikFromContext(context, "fixtures");
 			const errors = Object.entries(kivik.validateFixtures());
 			await kivik.close();
